+++ .storybook/config.js
@@ -2,7 +2,7 @@ import { addDecorator, addParameters, configure } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
 
 import { withThemeProvider } from './decorators'
-import parameters from './parameters'
+import parameters, { infoConfig } from './parameters'
 
 function loadStories() {
   const req = require.context('../stories', true, /\.*\.js$/)
@@ -11,5 +11,5 @@ function loadStories() {
 
 addParameters(parameters)
 configure(loadStories, module)
-addDecorator(withInfo)
+addDecorator(withInfo(infoConfig))
 addDecorator(withThemeProvider)