+++ src/components/BackToTop/index.js
@@ -38,4 +38,5 @@ const BackToTop = ({ className }) => {
   )
 }
 
+export { BackToTop }
 export default styled(BackToTop)``