+++ src/utils/isEmpty.js
@@ -1,1 +1,1 @@
-export default (array = []) => array.length < 1
+export default (obj = {}) => obj.length < 1 || Object.keys(obj).length < 1